System Scope Rework
System scope is now set in the releases list. The releases list now shows the tag of the releases. Storage of the system scope has been changed to be a list of app_ids instead of tags to better suit the specifications set by the professor. The setup.php file was reset to it's original version and does not contain any functionality.